Program 16 updated

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  * @range: fibonachi length
  * Return: nothing
  */
-void printFib(int range)
+void printFib(unsigned long int range)
 {
 	static unsigned long int first = 0, second = 1, sum, i;
